+++ aikau/src/test/resources/testApp/WEB-INF/classes/alfresco/site-webscripts/alfresco/upload/FileUploadService.get.js
@@ -18,7 +18,6 @@ model.jsonModel = {
                {
                   name: "alfresco/upload/UploadMonitor",
                   config: {
-                     useEllipsisForLongFilenames: true,
                      widgetsForSuccessfulActions: [
                         {
                            name: "alfresco/html/SVGImage",